TODO: finish parsing for the client request body. As well as remove as much EXIT(-1) as possible. Finally, fix errors, so that it compiles.

+++ include/Client.hpp
@@ -9,7 +9,7 @@ class Client {
 
 		std::string request_buffer;
 		
-        ssize_t		keep_alive_timer;
+        int		keep_alive_timer;
 
         ParseRequest    parser;
 		HttpRequest     req;
